Refactor package structure for API, context, metrics, resilience, and security
Moved controllers, services, and models into new 'api' subpackages for actioncard, hand, movement, and mygames. Consolidated context-related classes into 'context', metrics into 'metrics', resilience-related classes into 'resilience', and security/auth classes into 'security'. Updated package declarations and imports accordingly to improve project organization and maintainability.
