+++ src/main/java/ti4/spring/api/hand/GetHandResponse.java
@@ -1,0 +1,5 @@
+package ti4.spring.api.hand;
+
+import java.util.Set;
+
+public record GetHandResponse(Set<String> actionCards, Set<String> secretObjectives, Set<String> promissoryNotes) {}
