+++ src/main/java/ti4/spring/api/movement/MovementUnitCount.java
@@ -1,0 +1,9 @@
+package ti4.spring.api.movement;
+
+import java.util.List;
+
+/**
+ * Represents a single unit move with counts per state.
+ * counts: [regular, sustained]
+ */
+public record MovementUnitCount(String unitType, String colorID, List<Integer> counts) {}
