+++ src/main/java/ti4/spring/api/mygames/MyGameSummary.java
@@ -1,0 +1,3 @@
+package ti4.spring.api.mygames;
+
+public record MyGameSummary(String gameId, String faction, String color) {}
