+++ src/main/java/ti4/spring/context/WebConfiguration.java
@@ -1,0 +1,18 @@
+package ti4.spring.context;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@RequiredArgsConstructor
+@Configuration
+public class WebConfiguration implements WebMvcConfigurer {
+
+    private final GameLockAndRequestContextInterceptor gameLockAndRequestContextInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(gameLockAndRequestContextInterceptor).addPathPatterns("/api/game/{gameName}/**");
+    }
+}
