--- src/main/java/ti4/spring/controller/PingController.java
@@ -1,16 +1,0 @@
-package ti4.spring.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/public/ping")
-public class PingController {
-
-    @GetMapping
-    public ResponseEntity<String> ping() {
-        return ResponseEntity.ok("pong");
-    }
-}
