--- src/main/java/ti4/spring/exception/ServiceUnavailableException.java
@@ -1,12 +1,0 @@
-package ti4.spring.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.SERVICE_UNAVAILABLE)
-public class ServiceUnavailableException extends RuntimeException {
-
-    public ServiceUnavailableException() {
-        super("Unable to handle this request at this time");
-    }
-}
