--- src/main/java/ti4/spring/model/GetHandResponse.java
@@ -1,5 +1,0 @@
-package ti4.spring.model;
-
-import java.util.Set;
-
-public record GetHandResponse(Set<String> actionCards, Set<String> secretObjectives, Set<String> promissoryNotes) {}
