--- src/main/java/ti4/spring/model/MyGameSummary.java
@@ -1,3 +1,0 @@
-package ti4.spring.model;
-
-public record MyGameSummary(String gameId, String faction, String color) {}
