+++ src/main/java/ti4/spring/resilience/ServiceUnavailableException.java
@@ -1,0 +1,12 @@
+package ti4.spring.resilience;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.SERVICE_UNAVAILABLE)
+public class ServiceUnavailableException extends RuntimeException {
+
+    public ServiceUnavailableException() {
+        super("Unable to handle this request at this time");
+    }
+}
